requirements.txt
numpy==1.21.4
pickle-mixin==1.0.2
streamlit==1.2.0
streamlit-option-menu==0.3.2
scikit-learn==1.0.1
protobuf==3.20.1
Pillow  # 'PIL' is now 'Pillow' in modern versions
opencv-python
matplotlib
requests
tensorflow